Party System, Arena Management GUI, Command System Overhaul, and Bug Fixes
This update fixes many of the bugs that have been reported so far and adds two new features to SWR.
A basic party system, allowing friends to party up so that they can join the same games. Permissions are OP by default so must be giving to players.
A Arena GUI system that can be opened with the command /swmap arenas. This GUI displays information about each Arena for admins. Click on a arena, will open up a further menu that allows admins to change arena settings using the GUI and to even end matches.
The changes to the command system are quite intensive, so it is important that you look at the permission and command changes on the spigot overview page.